backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurocom"
version = "0.1.0"
description = "Real-time messaging social media platform"
dependencies = [
    "Django==5.2.3",
    "djangorestframework==3.15.2",
    "channels==4.0.0",
    "channels-redis==4.2.0",
    "daphne==4.1.2",
    "django-cleanup==9.0.0",
    "django-cors-headers==4.6.0",
    "django-polymorphic==4.1.0",
    "django-redis==5.4.0",
    "pillow==11.0.0",
    "psycopg2-binary==2.9.10",
    "python-decouple==3.8",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "redis==5.2.0",
    "python-json-logger==3.3.0"

]

[tool.setuptools]
package-dir = {"" = "src"}

[project.optional-dependencies]
dev = [
    "mypy==1.17.0",
    "django-stubs==5.2.2",
    "django-stubs-ext==5.2.2",
    "djangorestframework-stubs==3.15.1",
    "types-requests==2.32.0.20240914",
    "types-redis==4.6.0.20240726",
    "types-pillow==10.2.0.20240822",
    "types-PyYAML==6.0.12.20250516"
]


[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]

# Relaxed settings for Django projects
strict = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_calls = false
disallow_any_generics = false


warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = false  
check_untyped_defs = true
no_implicit_optional = true


exclude = [
    "build/",
    "dist/", 
    "venv/",
    "env/",
    ".venv/",
    "migrations/",
    ".*\\.egg-info/",
    "__pycache__/",
    "static/",
    "media/",
    "tests/"
]


[tool.django-stubs]
django_settings_module = "neurocom.settings"


[[tool.mypy.overrides]]
module = [
    "*.migrations.*",
    "migrations.*",
    "venv.*",
    "env.*",
    ".venv.*"
]
ignore_errors = true


[[tool.mypy.overrides]]
module = [
    "channels.*",
    "channels_redis.*", 
    "daphne.*",
    "django_cleanup.*",
    "corsheaders.*",
    "polymorphic.*",
    "django_redis.*",
    "decouple.*",
    "requests_oauthlib.*"
]
ignore_missing_imports = true


[[tool.mypy.overrides]]
module = "neurocom.*"
disallow_untyped_defs = false
warn_return_any = false
